import (OSCI internal) SystemC version 2.2.1.06nov08_beta
Summary: import (OSCI internal) SystemC version 2.2.1.06nov08_beta
Keywords:

Imported systemc-2.2.1.06nov08_beta
into Git repository

+++ src/sysc/communication/sc_port.cpp
@@ -41,6 +41,9 @@
 
 
 // $Log: sc_port.cpp,v $
+// Revision 1.2  2008/10/10 17:36:34  acg
+//  Andy Goodrich: update of copyright.
+//
 // Revision 1.1.1.1  2006/12/15 20:31:35  acg
 // SystemC 2.2
 //
@@ -188,7 +191,7 @@ struct sc_bind_info
 
 // constructor
 
-sc_bind_info::sc_bind_info( int max_size_, sc_port_policy policy_ )
+sc_bind_info::sc_bind_info( int    max_size_, sc_port_policy policy_ )
 : m_max_size( max_size_ ),
   m_policy( policy_ ),
   has_parent( false ),
@@ -208,7 +211,7 @@ sc_bind_info::~sc_bind_info()
 }
 
 
-int
+int   
 sc_bind_info::max_size() const
 {
     return m_max_size ? m_max_size : vec.size();
@@ -220,7 +223,7 @@ sc_bind_info::policy() const
     return m_policy;
 }
 
-int
+int   
 sc_bind_info::size() const
 {
     return vec.size();
@@ -429,7 +432,7 @@ sc_port_base::make_sensitive( sc_method_handle handle_,
 int
 sc_port_base::first_parent()
 {
-    for( int i = 0; i < m_bind_info->size(); ++ i ) {
+    for( int    i = 0; i < m_bind_info->size(); ++ i ) {
 	if( m_bind_info->vec[i]->parent != 0 ) {
 	    return i;
 	}
@@ -547,7 +550,7 @@ sc_port_base::complete_binding()
     if ( actual_binds > m_bind_info->max_size() )
     {
 	sprintf(msg_buffer, "%d binds exceeds maximum of %d allowed",
-	    actual_binds, m_bind_info->max_size() );
+	    (int)actual_binds, (int)m_bind_info->max_size() );
 	report_error( SC_ID_COMPLETE_BINDING_, msg_buffer );
     }
     switch ( m_bind_info->policy() )
@@ -560,7 +563,7 @@ sc_port_base::complete_binding()
       case SC_ALL_BOUND:
         if ( actual_binds < m_bind_info->max_size() || actual_binds < 1 ) {
 	    sprintf(msg_buffer, "%d actual binds is less than required %d",
-	        actual_binds, m_bind_info->max_size() ); 
+	        (int)actual_binds, (int)m_bind_info->max_size() ); 
             report_error( SC_ID_COMPLETE_BINDING_, msg_buffer );
         }
         break;
